refactor: charCount into its own component
